pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "sigpro"
version = "0.0.1"
dependencies = [
    "pyserial",
    "hid"
]
